Made changes according to the review
